alert

+++ components/FavoritesComponent.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+import * as Animatable from 'react-native-animatable';
 
 const mapStateToProps = state => {
     return {
@@ -28,20 +29,39 @@ class Favorites extends Component {
         const renderFavoriteItem = ({item}) => {
             const rightButton = [
                 {
-                    text: 'Delete',
+                    text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                    }
                 }
             ];
-
+            
             return (
                 <Swipeout right={rightButton} autoClose={true}>
-                    <ListItem
+                     <Animatable.View animation='fadeInRightBig' duration={2000}>
+                        <ListItem
                         title={item.name}
                         subtitle={item.description}
                         leftAvatar={{source: {uri: baseUrl + item.image}}}
                         onPress={() => navigate('CampsiteInfo', {campsiteId: item.id})}
-                    />
+                        />
+                    </Animatable.View>
                 </Swipeout>
             );
         };
@@ -59,6 +79,7 @@ class Favorites extends Component {
         }
         return (
             <FlatList
+                // filters out campsites with IDs matching the campsite IDs showing in the favorites array.
                 data={this.props.campsites.campsites.filter(
                     campsite => this.props.favorites.includes(campsite.id)
                 )}
